requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary>=2.9,<3.0
psycopg2>=2.9,<3.0
filelock==3.16.1
fonttools==4.54.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
mutagen==1.47.0
networkx==3.3
numpy==2.1.1
openai==1.37.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
psycopg2-binary==2.9.10
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-pptx==1.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==9.0.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
vboxapi==1.0
virtualenv==20.27.1
websockets==13.0
XlsxWriter==3.2.0
youtube-dl==2021.12.17
yt-dlp==2024.8.6
